Implement get-compact-range using RFC 6962 methods
This is a proof of concept change demonstrating that it is possible to
obtain arbitrary compact ranges from a Merkle tree log that restricts
itself only to endpoints represented in RFC 6962, in constant time
interaction complexity.

Specifically, it is possible to obtain comact range [begin, end) by
calling "get consistency proof" endpoints <= 2 times for carefully
crafted tree sizes. In a few cases where it is impossible to get certain
hashes, this approach falls back to calling the "get entries" endpoint 1
time to obtain between 1-3 entries and reconstruct the compact range.

Overall, the interaction with the log is limited by 2 calls, and each
call is limited in size.