compact: Visit the seed node in the Append method (#20)
Most uses of compact range invoke the visitor for the leaf node before
the Append call. This change makes it the default behaviour of Append.

The asymmetry between Append and AppendRange can be explained by the
fact that Append is usually called when the new leaf hasn't been
processed yet, whereas AppendRange is usually called when all the nodes
of the compact ranges (as well as nodes below them) have been processed.