proof: Avoid all reallocations (#23)
This change introduces pre-allocation into the Proof generation types, thereby
reducing the total number of slice allocations to one. A new helper for computing
the compact range size compact.RangeSize is introduced, and the signature of
the previous compact.RangeNodes now requires providing a slice to append the
result to.