proof: Allow single child for the ephemeral node (#22)
The ephemeral node is the node that does not yet exist, and is therefore replaced
by its "children". It can happen that there is only one "child" node representing the
ephemeral node (i.e. when the subtree that it covers is perfect). If this is the case,
we still want to expose this signal to the user of Proof type, which can then choose
whether to treat this node as ephemeral or not.